requirements.txt
annotated-types==0.6.0
bert-extractive-summarizer==0.10.1
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
filelock==3.13.1
Flask==3.0.0
fsspec==2023.10.0
gensim==4.3.2
huggingface-hub==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
packaging==23.2
preshed==3.0.9
pydantic==2.5.0
pydantic_core==2.14.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.35.1
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
